Revert "Refactor ANT pattern logic (#716)"
This reverts commit 3ac03a5ba76abefea926eceea3ff51c12b2f487a.
